data_format_translation_servers/python/pyproject.toml
[tool.poetry]
name = "data-format-translation-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
fastapi = "^0.85.0"
uvicorn = "^0.18.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
